[Inductor][Triton] Support TMA before strict 3.4 cutoff (#159777)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/159777

Inductor's 3.4 Triton release is the most common used variant of Triton, but if someone is working with an alternative version of Triton this may not match. This moves the version check from 3.4 Triton to any variant that has support for the TMA APIs.

Test Plan:
Relying on CI. Should be a NFC.

Rollback Plan:

Reviewed By: davidberard98

Differential Revision: D79378792